refactor(harness): group the renderer views under one directory
